# apps/shared/fastapi_docs.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "coding-agent backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/workspaces": {
      "get": {
        "summary": "List Workspaces",
        "description": "返回已登记的工作区列表。\n\n参数:\n    workspace_service: 通过依赖注入的工作区 service。\n\n返回:\n    ``WorkspaceResponse`` 列表。\n\n异常:\n    无。\n\n副作用:\n    无。",
        "operationId": "list_workspaces_workspaces_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WorkspaceResponse"
                  },
                  "type": "array",
                  "title": "Response List Workspaces Workspaces Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Workspace",
        "description": "创建一个本地工作区。\n\n参数:\n    payload: 包含 name 与 root_path 的请求体。\n    workspace_service: 通过依赖注入的工作区 service。\n\n返回:\n    创建后的 ``WorkspaceResponse``。\n\n异常:\n    HTTPException: 当输入非法时抛出。\n\n副作用:\n    在存储中创建工作区。",
        "operationId": "create_workspace_workspaces_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkspaceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/workspaces/{workspace_id}": {
      "delete": {
        "summary": "Delete Workspace",
        "description": "删除工作区及其下游任务、轮次与运行记录。\n\n参数:\n    workspace_id: 来自路由的工作区标识。\n    workspace_service: 通过依赖注入的工作区 service。\n\n返回:\n    删除结果 ``DeleteWorkspaceResponse``。\n\n异常:\n    HTTPException: 当工作区不存在时抛出。\n\n副作用:\n    级联删除工作区下的任务、轮次、事件、步骤与 durable run。",
        "operationId": "delete_workspace_workspaces__workspace_id__delete",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Workspace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteWorkspaceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/workspaces/{workspace_id}/tasks": {
      "get": {
        "summary": "List Workspace Tasks",
        "description": "返回工作区下的任务列表。\n\n参数:\n    workspace_id: 来自路由的工作区标识。\n    task_service: 通过依赖注入的任务 service。\n\n返回:\n    ``TaskResponse`` 列表。\n\n异常:\n    HTTPException: 当工作区不存在时抛出。\n\n副作用:\n    无。",
        "operationId": "list_workspace_tasks_workspaces__workspace_id__tasks_get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Workspace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskResponse"
                  },
                  "title": "Response List Workspace Tasks Workspaces  Workspace Id  Tasks Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Workspace Task",
        "description": "在工作区下创建任务容器（不含首轮次）。\n\n任务创建与首轮次创建已解耦：本端点只创建 task 容器，首轮次由调用方（前端）显式\n调 ``POST /tasks/{task_id}/turns`` 创建。这样 task 维度不再耦合 agent / 模型 /\n附件等仅属于 turn 的字段，agent 由 turn 维度承载。\n\n参数:\n    payload: 创建任务请求体，含任务文本与 workspace。\n    task_service: 任务 service，用于创建任务容器。\n\n返回:\n    新建任务记录（含 ``task_id``）。调用方随后创建首 turn 并连接\n    ``/turns/{turn_id}/stream`` 驱动执行。\n\n异常:\n    HTTPException: workspace 不存在时为 404；输入非法时为 400。\n\n副作用:\n    向 ``tasks`` 插入一行任务容器；不创建轮次。",
        "operationId": "create_workspace_task_workspaces__workspace_id__tasks_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/workspaces/{workspace_id}/events/stream": {
      "get": {
        "summary": "Stream Workspace Events",
        "description": "以 SSE 流式返回 workspace 状态事件。\n\n通用 workspace 状态事件通道：当前承载创建时的准备进度（preparing/ready/degraded），\n后续可扩展其他 workspace 状态事件。客户端须先调用本端点建立订阅，再触发 prepare，\n避免错过 preparing 事件。\n\n参数:\n    workspace_id: 来自路由的 workspace 标识。\n    event_bus: workspace 级状态事件总线。\n\n返回:\n    text/event-stream 的 StreamingResponse；终态事件（ready/degraded）后结束流。\n\n异常:\n    无（订阅缺失时流自然结束）。\n\n副作用:\n    注册并最终移除一条 workspace 状态事件订阅。",
        "operationId": "stream_workspace_events_workspaces__workspace_id__events_stream_get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Workspace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/workspaces/{workspace_id}/events/prepare": {
      "post": {
        "summary": "Prepare Workspace",
        "description": "同步触发一次 workspace 准备（prepare）。\n\n当前准备动作是 CodeGraph 索引就绪；该端点走通用 workspace 状态事件通道，后续可扩展\n其他 workspace 状态事件的准备动作。客户端应先连 ``/events/stream`` 建立订阅再触发本端点。\n\n参数:\n    workspace_id: 来自路由的 workspace 标识。\n    workspace_service: 工作区 service（用于取 root_path 与 404 守卫）。\n    event_service: workspace 状态事件编排 service；Kernel 不可用时为 None（降级）。\n    event_bus: workspace 级状态事件总线（用于 Kernel 不可用时主动发降级终态）。\n\n返回:\n    WorkspacePrepareResponse，含就绪状态与动作摘要。\n\n异常:\n    HTTPException: 当 workspace 不存在（404）时抛出。\n\n副作用:\n    发布 preparing/ready/degraded 工作状态事件到总线。\n\n设计要点:\n    Kernel 不可用时 event_service 为 None，属于设计明确的正常降级场景。此时不仅\n    HTTP 返回 unavailable，还**主动 emit 一条 WORKSPACE_DEGRADED 事件**到总线——\n    因为前端先连 SSE 再 POST prepare，若只靠 HTTP 响应而 SSE 订阅者未收到终态事件，\n    前端会永远停在 preparing（独立审查暴露的时序 bug 的另一半）。",
        "operationId": "prepare_workspace_workspaces__workspace_id__events_prepare_post",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Workspace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspacePrepareResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateTaskRequest": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "workspace_id": {
            "type": "integer",
            "title": "Workspace Id"
          }
        },
        "type": "object",
        "required": [
          "text",
          "workspace_id"
        ],
        "title": "CreateTaskRequest",
        "description": "校验任务创建请求体（仅建 task 容器，不含首轮次）。\n\n任务创建与首轮次创建已解耦：本请求只携带建 task 容器所需的最小字段\n（``text`` 派生标题、``workspace_id`` 归属工作区）。agent / 模型 / 附件等\n属于 turn 维度的字段由 ``CreateTurnRequest`` 承载，调用方在创建首 turn 时单独传递。\n\n参数:\n    text: 非空的纯文本任务输入。\n    workspace_id: 工作区标识。\n\n返回:\n    Pydantic 请求模型。\n\n异常:\n    ValueError: 当 ``text`` 为空白时抛出。\n\n副作用:\n    无。"
      },
...
      "CreateWorkspaceRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "root_path": {
            "type": "string",
            "title": "Root Path"
          }
        },
        "type": "object",
        "required": [
          "name",
          "root_path"
        ],
        "title": "CreateWorkspaceRequest",
        "description": "校验工作区创建请求体。\n\n参数:\n    name: 非空的工作区名称。\n    root_path: 非空的工作区本地路径。\n\n返回:\n    Pydantic 请求模型。\n\n异常:\n    ValueError: 当名称或路径为空白时抛出。\n\n副作用:\n    无。"
      },
...
      "DeleteWorkspaceResponse": {
        "properties": {
          "workspace_id": {
            "type": "integer",
            "title": "Workspace Id"
          },
          "deleted": {
            "type": "boolean",
            "title": "Deleted"
          }
        },
        "type": "object",
        "required": [
          "workspace_id",
          "deleted"
        ],
        "title": "DeleteWorkspaceResponse",
        "description": "校验并序列化工作区删除结果响应。\n\n参数:\n    workspace_id: 被删除的工作区标识。\n    deleted: 是否删除成功。\n\n返回:\n    Pydantic 响应模型。\n\n异常:\n    无。\n\n副作用:\n    无。"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TaskResponse": {
        "properties": {
          "task_id": {
            "type": "integer",
            "title": "Task Id"
          },
          "workspace_id": {
            "type": "integer",
            "title": "Workspace Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "execution_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution Status"
          },
          "task_type": {
            "type": "string",
            "title": "Task Type",
            "default": "user"
          },
          "parent_task_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Task Id"
          },
          "parent_turn_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Turn Id"
          },
          "delegation_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delegation Id"
          },
          "context_usage_used": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Usage Used"
          },
          "context_window_total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Window Total"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "workspace_id",
          "title",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "TaskResponse",
        "description": "校验并序列化任务状态响应（与 ``TaskRecord.to_dict()`` 对齐）。\n\n任务不再绑定 agent：本响应不含 agent_id，agent 维度由 turn 维度承载。\n\n参数:\n    task_id: 任务标识。\n    workspace_id: 所属工作区标识。\n    title: 任务标题。\n    status: 任务生命周期状态。\n    execution_status: 派生执行状态，可能为 None。\n    task_type: 任务类型，``\"user\"`` 为用户创建，``\"delegation\"`` 为委派子任务。\n    parent_task_id: 父任务标识，仅委派子任务有值。\n    parent_turn_id: 父轮次标识，仅委派子任务有值。\n    delegation_id: 所属委派标识，仅委派子任务有值。\n    context_usage_used: 最近一次上下文窗口已用 token（运行时回写，可能为 None）。\n    context_window_total: 该任务模型的上下文窗口上限 token（由 resolve_context_window\n        动态计算，可能为 None）。\n    created_at: 创建时间文本。\n    updated_at: 更新时间文本。\n\n返回:\n    Pydantic 响应模型。\n\n异常:\n    无。\n\n副作用:\n    无。"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WorkspacePrepareResponse": {
        "properties": {
          "workspace_id": {
            "type": "integer",
            "title": "Workspace Id"
          },
          "ready": {
            "type": "boolean",
            "title": "Ready"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "action_taken": {
            "type": "string",
            "title": "Action Taken",
            "default": "none"
          },
          "files_changed": {
            "type": "integer",
            "title": "Files Changed",
            "default": 0
          },
          "duration_ms": {
            "type": "integer",
            "title": "Duration Ms",
            "default": 0
          },
          "degraded_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Degraded Reason"
          }
        },
        "type": "object",
        "required": [
          "workspace_id",
          "ready",
          "state"
        ],
        "title": "WorkspacePrepareResponse",
        "description": "POST /workspaces/{workspace_id}/events/prepare 的响应。\n\n触发一次 workspace 准备（当前为 CodeGraph 索引就绪）并立即返回；\n准备进度经 ``GET /workspaces/{workspace_id}/events/stream`` SSE 推送。\n\n属性:\n    workspace_id: 触发准备的 workspace 标识。\n    ready: 即时快照中 Kernel 是否已就绪（True / False）。\n    state: 即时状态快照，取值 ``ready`` / ``failed`` / ``unavailable`` /\n        ``unreachable`` / ``timeout``；任何非 ``ready`` 值均表示未就绪，前端应\n        降级到文件搜索但任务可继续。同步阻塞路径下通常为 ``accepted`` 初值，\n        终态以 SSE 事件为准。\n    action_taken: 采取的动作（``init`` / ``sync`` / ``none``）。\n    files_changed: 索引变更文件数（ready 时有效）。\n    duration_ms: 准备耗时（毫秒）。\n    degraded_reason: 降级原因（非 ready 时有效；脱敏，不记录 secret）。"
      },
      "WorkspaceResponse": {
        "properties": {
          "workspace_id": {
            "type": "integer",
            "title": "Workspace Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "root_path": {
            "type": "string",
            "title": "Root Path"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "workspace_id",
          "name",
          "root_path",
          "created_at",
          "updated_at"
        ],
        "title": "WorkspaceResponse",
        "description": "校验并序列化工作区状态响应（与 ``WorkspaceRecord.to_dict()`` 对齐）。\n\n参数:\n    workspace_id: 工作区标识。\n    name: 工作区名称。\n    root_path: 工作区根目录路径。\n    created_at: 创建时间文本。\n    updated_at: 更新时间文本。\n\n返回:\n    Pydantic 响应模型。\n\n异常:\n    无。\n\n副作用:\n    无。"
      }
    }
  }
}